+++ packages/auth/src/config.ts
@@ -1,0 +1,31 @@
+import type { DefaultSession, NextAuthConfig } from "next-auth";
+import { DrizzleAdapter } from "@auth/drizzle-adapter";
+import Discord from "next-auth/providers/discord";
+
+import { db, tableCreator } from "@acme/db";
+
+declare module "next-auth" {
+  interface Session {
+    user: {
+      id: string;
+    } & DefaultSession["user"];
+  }
+}
+
+export const authConfig = {
+  adapter: DrizzleAdapter(db, tableCreator),
+  providers: [Discord],
+  callbacks: {
+    session: (opts) => {
+      if (!("user" in opts)) throw "unreachable with session strategy";
+
+      return {
+        ...opts.session,
+        user: {
+          ...opts.session.user,
+          id: opts.user.id,
+        },
+      };
+    },
+  },
+} satisfies NextAuthConfig;
