+++ packages/auth/src/index.ts
@@ -1,48 +1,14 @@
-import type { DefaultSession } from "next-auth";
-import { cache } from "react";
-import { DrizzleAdapter } from "@auth/drizzle-adapter";
 import NextAuth from "next-auth";
-import Discord from "next-auth/providers/discord";
 
-import { db, tableCreator } from "@acme/db";
+import { authConfig } from "./config";
 
 export type { Session } from "next-auth";
 
-declare module "next-auth" {
-  interface Session {
-    user: {
-      id: string;
-    } & DefaultSession["user"];
-  }
-}
-
 const {
   handlers: { GET, POST },
-  auth: defaultAuth,
+  auth,
   signIn,
   signOut,
-} = NextAuth({
-  adapter: DrizzleAdapter(db, tableCreator),
-  providers: [Discord],
-  callbacks: {
-    session: (opts) => {
-      if (!("user" in opts)) throw "unreachable with session strategy";
+} = NextAuth(authConfig);
 
-      return {
-        ...opts.session,
-        user: {
-          ...opts.session.user,
-          id: opts.user.id,
-        },
-      };
-    },
-  },
-});
-
-export { GET, POST, signIn, signOut };
-
-/**
- * This is the main way to get session data for your RSCs.
- * This will de-duplicate all calls to next-auth's default `auth()` function and only call it once per request across all components
- */
-export const auth = cache(defaultAuth);
+export { GET, POST, auth, signIn, signOut };
